+++ features/kategori/entities.go
@@ -7,7 +7,7 @@ import (
 type Kategori struct {
 	gorm.Model
 
-	ID        uint    `gorm:"primaryKey"`
+	ID        uint64  `gorm:"primaryKey"`
 	Name      string  `gorm:"type:varchar(255);not null"`
 	Deskripsi *string `gorm:"type:text"`
 }
