+++ features/log_audit/interfaces.go
@@ -9,17 +9,17 @@ import (
 type Repository interface {
 	GetAll(page, size int) ([]Log_audit, int64, error)
 	Insert(newLog_audit Log_audit) error
-	SelectByID(log_auditID uint) (*Log_audit, error)
+	SelectByID(log_auditID uint64) (*Log_audit, error)
 	Update(log_audit Log_audit) error
-	DeleteByID(log_auditID uint) error
+	DeleteByID(log_auditID uint64) error
 }
 
 type Usecase interface {
 	FindAll(page, size int) ([]dtos.ResLog_audit, int64, error)
-	FindByID(log_auditID uint) (*dtos.ResLog_audit, error)
+	FindByID(log_auditID uint64) (*dtos.ResLog_audit, error)
 	Create(newLog_audit dtos.InputLog_audit) error
-	Modify(log_auditData dtos.InputLog_audit, log_auditID uint) error
-	Remove(log_auditID uint) error
+	Modify(log_auditData dtos.InputLog_audit, log_auditID uint64) error
+	Remove(log_auditID uint64) error
 }
 
 type Handler interface {
