+++ features/master_data/entities.go
@@ -7,25 +7,25 @@ import (
 type Master_data struct {
 	gorm.Model
 
-	ID   uint   `gorm:"primaryKey"`
+	ID   uint64 `gorm:"primaryKey"`
 	Name string `gorm:"type:varchar(255)"`
 }
 type JenisPembayaran struct {
 	gorm.Model
 
-	ID   uint   `gorm:"primaryKey"`
+	ID   uint64 `gorm:"primaryKey"`
 	Nama string `gorm:"type:varchar(255)"`
 }
 type TipeTransaksi struct {
 	gorm.Model
 
-	ID   uint   `gorm:"primaryKey"`
+	ID   uint64 `gorm:"primaryKey"`
 	Nama string `gorm:"type:varchar(255)"`
 }
 type StatusTransaksi struct {
 	gorm.Model
 
-	ID   uint   `gorm:"primaryKey"`
+	ID   uint64 `gorm:"primaryKey"`
 	Nama string `gorm:"type:varchar(255)"`
 }
